+++ tests/test_dataset.py
@@ -13,7 +13,14 @@ from shutil import rmtree, copytree
 
 from wkw.wkw import WKWException
 
-from wkcuber.api.dataset import Dataset, DatasetViewConfiguration
+from wkcuber.api.properties import (
+    dataset_converter,
+    DatasetProperties,
+    _snake_to_camel_case,
+    SegmentationLayerProperties,
+    DatasetViewConfiguration,
+)
+from wkcuber.api.dataset import Dataset, PROPERTIES_FILE_NAME
 from wkcuber.api.bounding_box import BoundingBox
 from os import makedirs
 
@@ -23,13 +30,10 @@ from wkcuber.api.layer import (
     LayerViewConfiguration,
 )
 from wkcuber.api.mag_view import MagView
-from wkcuber.api.properties.dataset_properties import Properties
-from wkcuber.api.properties.layer_properties import SegmentationLayerProperties
-from wkcuber.api.properties.resolution_properties import Resolution
 from wkcuber.api.view import View
 from wkcuber.compress import compress_mag_inplace
 from wkcuber.mag import Mag
-from wkcuber.utils import get_executor_for_args, named_partial, _snake_to_camel_case
+from wkcuber.utils import get_executor_for_args, named_partial
 
 TESTDATA_DIR = Path("testdata")
 TESTOUTPUT_DIR = Path("testoutput")
@@ -150,8 +154,8 @@ def test_create_dataset_with_layer_and_mag() -> None:
     assert (TESTOUTPUT_DIR / "wk_dataset" / "color" / "1").exists()
     assert (TESTOUTPUT_DIR / "wk_dataset" / "color" / "2-2-1").exists()
 
-    assert len(ds.properties.data_layers) == 1
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 2
+    assert len(ds.layers) == 1
+    assert len(ds.get_layer("color").mags) == 2
 
 
 def test_create_dataset_with_explicit_header_fields() -> None:
@@ -168,31 +172,28 @@ def test_create_dataset_with_explicit_header_fields() -> None:
     assert (TESTOUTPUT_DIR / "wk_dataset_advanced" / "color" / "1").exists()
     assert (TESTOUTPUT_DIR / "wk_dataset_advanced" / "color" / "2-2-1").exists()
 
-    assert len(ds.properties.data_layers) == 1
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 2
+    assert len(ds.layers) == 1
+    assert len(ds.get_layer("color").mags) == 2
 
-    assert ds.properties.data_layers["color"].element_class == "uint48"
+    assert ds.get_layer("color").dtype_per_channel == "uint16"
+    assert ds.get_layer("color")._properties.element_class == "uint48"
+    assert ds.get_layer("color").get_mag(1).header.block_len == 64
+    assert ds.get_layer("color").get_mag(1).header.file_len == 64
+    assert ds.get_layer("color").get_mag(1)._properties.cube_length == 64 * 64
     assert (
-        cast(
-            Resolution, ds.properties.data_layers["color"].wkw_magnifications[0]
-        ).cube_length
-        == 64 * 64
-    )  # mag "1"
-    assert ds.properties.data_layers["color"].wkw_magnifications[0].mag == Mag("1")
+        ds.get_layer("color").get_mag("2-2-1").header.block_len == 32
+    )  # defaults are used
     assert (
-        cast(
-            Resolution, ds.properties.data_layers["color"].wkw_magnifications[1]
-        ).cube_length
-        == 32 * 32
-    )  # mag "2-2-1" (defaults are used)
-    assert ds.properties.data_layers["color"].wkw_magnifications[1].mag == Mag("2-2-1")
+        ds.get_layer("color").get_mag("2-2-1").header.file_len == 32
+    )  # defaults are used
+    assert ds.get_layer("color").get_mag("2-2-1")._properties.cube_length == 32 * 32
 
 
 def test_open_dataset() -> None:
     ds = Dataset(TESTDATA_DIR / "simple_wk_dataset")
 
-    assert len(ds.properties.data_layers) == 1
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 1
+    assert len(ds.layers) == 1
+    assert len(ds.get_layer("color").mags) == 1
 
 
 def test_modify_existing_dataset() -> None:
@@ -299,11 +300,11 @@ def test_mag_view_write_out_of_bounds() -> None:
     ds = Dataset(new_dataset_path)
     mag_view = ds.get_layer("color").get_mag("1")
 
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (24, 24, 24)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 24)
     mag_view.write(
         np.zeros((3, 1, 1, 48), dtype=np.uint8)
     )  # this is bigger than the bounding_box
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (24, 24, 48)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 48)
 
 
 def test_mag_view_write_out_of_bounds_mag2() -> None:
@@ -315,13 +316,13 @@ def test_mag_view_write_out_of_bounds_mag2() -> None:
     ds = Dataset(new_dataset_path)
     mag_view = ds.get_layer("color").get_or_add_mag("2-2-1")
 
-    assert ds.properties.data_layers["color"].get_bounding_box_offset() == (0, 0, 0)
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (24, 24, 24)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 24)
     mag_view.write(
         np.zeros((3, 50, 1, 48), dtype=np.uint8), (10, 10, 10)
     )  # this is bigger than the bounding_box
-    assert ds.properties.data_layers["color"].get_bounding_box_offset() == (0, 0, 0)
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (120, 24, 58)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (120, 24, 58)
 
 
 def test_update_new_bounding_box_offset() -> None:
@@ -330,21 +331,21 @@ def test_update_new_bounding_box_offset() -> None:
     ds = Dataset.create(TESTOUTPUT_DIR / "wk_dataset", scale=(1, 1, 1))
     mag = ds.add_layer("color", LayerCategories.COLOR_TYPE).add_mag("1")
 
-    assert ds.properties.data_layers["color"].bounding_box["topLeft"] == (-1, -1, -1)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
 
     np.random.seed(1234)
     write_data = (np.random.rand(10, 10, 10) * 255).astype(np.uint8)
     mag.write(
         write_data, offset=(10, 10, 10)
     )  # the write method of MagDataset does always use the relative offset to (0, 0, 0)
-    assert ds.properties.data_layers["color"].get_bounding_box_offset() == (10, 10, 10)
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (10, 10, 10)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (10, 10, 10)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (10, 10, 10)
 
     mag.write(
         write_data, offset=(5, 5, 20)
     )  # the write method of MagDataset does always use the relative offset to (0, 0, 0)
-    assert ds.properties.data_layers["color"].get_bounding_box_offset() == (5, 5, 10)
-    assert ds.properties.data_layers["color"].get_bounding_box_size() == (15, 15, 20)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (5, 5, 10)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (15, 15, 20)
 
 
 def test_write_multi_channel_uint8() -> None:
@@ -406,42 +407,6 @@ def test_read_padded_data() -> None:
 
     assert data.shape == (3, 10, 10, 10)
     assert np.array_equal(data, np.zeros((3, 10, 10, 10)))
-
-
-def test_read_and_write_of_properties() -> None:
-    destination_path = TESTOUTPUT_DIR / "read_write_properties"
-    delete_dir(destination_path)
-    source_file_name = TESTDATA_DIR / "simple_wk_dataset" / "datasource-properties.json"
-    destination_file_name = destination_path / "datasource-properties.json"
-
-    imported_properties = Properties._from_json(source_file_name)
-    imported_properties._path = destination_file_name
-    makedirs(destination_path)
-    imported_properties._export_as_json()
-
-    with open(source_file_name) as source_stream:
-        source_data = json.load(source_stream)
-        with open(destination_file_name) as destination_stream:
-            destination_data = json.load(destination_stream)
-            assert source_data == destination_data
-
-
-def test_read_and_write_of_view_configuration() -> None:
-    destination_path = TESTOUTPUT_DIR / "read_write_view_configuration"
-    delete_dir(destination_path)
-    source_file_name = TESTDATA_DIR / "simple_wk_dataset" / "datasource-properties.json"
-    destination_file_name = destination_path / "datasource-properties.json"
-
-    imported_properties = Properties._from_json(source_file_name)
-    imported_properties._path = destination_file_name
-    makedirs(destination_path)
-    imported_properties._export_as_json()
-
-    with open(source_file_name) as source_stream:
-        source_data = json.load(source_stream)
-        with open(destination_file_name) as destination_stream:
-            destination_data = json.load(destination_stream)
-            assert source_data == destination_data
 
 
 def test_num_channel_mismatch_assertion() -> None:
@@ -532,14 +497,14 @@ def test_open_dataset_without_num_channels_in_properties() -> None:
         assert data["dataLayers"][0].get("num_channels") is None
 
     ds = Dataset(TESTOUTPUT_DIR / "old_wk_dataset")
-    assert ds.properties.data_layers["color"].num_channels == 1
-    ds.properties._export_as_json()
+    assert ds.get_layer("color").num_channels == 1
+    ds._export_as_json()
 
     with open(
         TESTOUTPUT_DIR / "old_wk_dataset" / "datasource-properties.json"
     ) as datasource_properties:
         data = json.load(datasource_properties)
-        assert data["dataLayers"][0].get("num_channels") == 1
+        assert data["dataLayers"][0].get("numChannels") == 1
 
 
 def test_largest_segment_id_requirement() -> None:
@@ -559,50 +524,59 @@ def test_largest_segment_id_requirement() -> None:
 
     ds = Dataset(path)
     assert (
-        cast(
-            SegmentationLayerProperties, ds.properties.data_layers["segmentation"]
-        ).largest_segment_id
+        cast(SegmentationLayer, ds.get_layer("segmentation")).largest_segment_id
         == largest_segment_id
     )
 
 
 def test_properties_with_segmentation() -> None:
-    input_json_path = (
-        TESTDATA_DIR / "complex_property_ds" / "datasource-properties.json"
+    delete_dir(TESTOUTPUT_DIR / "complex_property_ds")
+    copytree(
+        TESTDATA_DIR / "complex_property_ds",
+        TESTOUTPUT_DIR / "complex_property_ds",
     )
-    output_json_path = (
-        TESTOUTPUT_DIR / "complex_property_ds" / "datasource-properties.json"
-    )
-    properties = Properties._from_json(input_json_path)
 
-    # the attributes 'largest_segment_id' and 'mappings' only exist if it is a SegmentationLayer
-    segmentation_layer = cast(
-        SegmentationLayerProperties, properties.data_layers["segmentation"]
-    )
-    assert segmentation_layer.largest_segment_id == 1000000000
-    assert segmentation_layer.mappings == [
-        "larger5um1",
-        "axons",
-        "astrocyte-ge-7",
-        "astrocyte",
-        "mitochondria",
-        "astrocyte-full",
-    ]
+    input_path = TESTOUTPUT_DIR / "complex_property_ds"
 
-    # export the json under a new name
-    makedirs(dirname(output_json_path), exist_ok=True)
-    properties._path = output_json_path
-    properties._export_as_json()
+    with open(input_path / "datasource-properties.json", "r") as f:
+        data = json.load(f)
+        ds_properties = dataset_converter.structure(data, DatasetProperties)
+
+        # the attributes 'largest_segment_id' and 'mappings' only exist if it is a SegmentationLayer
+        segmentation_layer = cast(
+            SegmentationLayerProperties,
+            [l for l in ds_properties.data_layers if l.name == "segmentation"][0],
+        )
+        assert segmentation_layer.largest_segment_id == 1000000000
+        assert segmentation_layer.mappings == [
+            "larger5um1",
+            "axons",
+            "astrocyte-ge-7",
+            "astrocyte",
+            "mitochondria",
+            "astrocyte-full",
+        ]
+
+    with open(input_path / "datasource-properties.json", "w") as f:
+        # Update the properties on disk (without changing the data)
+        json.dump(
+            dataset_converter.unstructure(ds_properties),
+            f,
+            indent=4,
+        )
 
     # validate if contents match
-    with open(input_json_path) as input_properties:
+    with open(
+        TESTDATA_DIR / "complex_property_ds" / "datasource-properties.json"
+    ) as input_properties:
         input_data = json.load(input_properties)
 
-        with open(output_json_path) as output_properties:
+        with open(input_path / "datasource-properties.json", "r") as output_properties:
             output_data = json.load(output_properties)
             for layer in output_data["dataLayers"]:
                 # remove the num_channels because they are not part of the original json
-                del layer["num_channels"]
+                if "numChannels" in layer:
+                    del layer["numChannels"]
 
             assert input_data == output_data
 
@@ -666,8 +640,8 @@ def test_view_write_without_open() -> None:
 
     ds = Dataset.create(ds_path, scale=(1, 1, 1))
     layer = ds.add_layer("color", LayerCategories.COLOR_TYPE)
-    layer.set_bounding_box(
-        offset=(0, 0, 0), size=(64, 64, 64)
+    layer.bounding_box = BoundingBox(
+        (0, 0, 0), (64, 64, 64)
     )  # This newly created dataset would otherwise have a "empty" bounding box
     mag = layer.add_mag("1")
 
@@ -730,37 +704,40 @@ def test_changing_layer_bounding_box() -> None:
     layer = ds.get_layer("color")
     mag = layer.get_mag("1")
 
-    bbox_size = ds.properties.data_layers["color"].get_bounding_box_size()
-    assert bbox_size == (24, 24, 24)
+    bbox_size = ds.get_layer("color").bounding_box.size
+    assert tuple(bbox_size) == (24, 24, 24)
     original_data = mag.read(size=bbox_size)
     assert original_data.shape == (3, 24, 24, 24)
 
-    layer.set_bounding_box_size((12, 12, 10))  # decrease bounding box
+    old_bbox = layer.bounding_box
+    old_bbox.size = np.array([12, 12, 10])
+    layer.bounding_box = old_bbox  # decrease bounding box
 
-    bbox_size = ds.properties.data_layers["color"].get_bounding_box_size()
-    assert bbox_size == (12, 12, 10)
+    bbox_size = ds.get_layer("color").bounding_box.size
+    assert tuple(bbox_size) == (12, 12, 10)
     less_data = mag.read(size=bbox_size)
     assert less_data.shape == (3, 12, 12, 10)
     assert np.array_equal(original_data[:, :12, :12, :10], less_data)
 
-    layer.set_bounding_box_size((36, 48, 60))  # increase the bounding box
+    old_bbox.size = np.array([36, 48, 60])
+    layer.bounding_box = old_bbox  # increase the bounding box
 
-    bbox_size = ds.properties.data_layers["color"].get_bounding_box_size()
-    assert bbox_size == (36, 48, 60)
+    bbox_size = ds.get_layer("color").bounding_box.size
+    assert tuple(bbox_size) == (36, 48, 60)
     more_data = mag.read(size=bbox_size)
     assert more_data.shape == (3, 36, 48, 60)
     assert np.array_equal(more_data[:, :24, :24, :24], original_data)
 
-    assert ds.properties.data_layers["color"].get_bounding_box_offset() == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
 
     # Move the offset from (0, 0, 0) to (10, 10, 0)
     # Note that the bottom right coordinate of the dataset is still at (24, 24, 24)
-    layer.set_bounding_box(offset=(10, 10, 0), size=(14, 14, 24))
+    layer.bounding_box = BoundingBox((10, 10, 0), (14, 14, 24))
 
-    new_bbox_offset = ds.properties.data_layers["color"].get_bounding_box_offset()
-    new_bbox_size = ds.properties.data_layers["color"].get_bounding_box_size()
-    assert new_bbox_offset == (10, 10, 0)
-    assert new_bbox_size == (14, 14, 24)
+    new_bbox_offset = ds.get_layer("color").bounding_box.topleft
+    new_bbox_size = ds.get_layer("color").bounding_box.size
+    assert tuple(new_bbox_offset) == (10, 10, 0)
+    assert tuple(new_bbox_size) == (14, 14, 24)
     # Note that even though the offset was changed (in the properties), the offset of 'mag.read()'
     # still refers to the absolute position (relative to (0, 0, 0)).
     # The default offset is (0, 0, 0). Since the bottom right did not change, the read data equals 'original_data'.
@@ -772,7 +749,7 @@ def test_changing_layer_bounding_box() -> None:
 
     # resetting the offset to (0, 0, 0)
     # Note that the size did not change. Therefore, the new bottom right is now at (14, 14, 24)
-    layer.set_bounding_box_offset((0, 0, 0))
+    layer.bounding_box = BoundingBox((0, 0, 0), new_bbox_size)
     new_data = mag.read()
     assert new_data.shape == (3, 14, 14, 24)
     assert np.array_equal(original_data[:, :14, :14, :], new_data)
@@ -945,16 +922,16 @@ def test_adding_layer_with_valid_dtype_per_layer() -> None:
     ds = Dataset(
         TESTOUTPUT_DIR / "valid_dtype"
     )  # reopen the dataset to check if the data is read from the properties correctly
-    assert ds.properties.data_layers["color1"].element_class == "uint24"
-    assert ds.properties.data_layers["color2"].element_class == "uint8"
-    assert ds.properties.data_layers["color3"].element_class == "uint24"
-    assert ds.properties.data_layers["color4"].element_class == "uint24"
+    assert ds.get_layer("color1").dtype_per_layer == "uint24"
+    assert ds.get_layer("color2").dtype_per_layer == "uint8"
+    assert ds.get_layer("color3").dtype_per_layer == "uint24"
+    assert ds.get_layer("color4").dtype_per_layer == "uint24"
     # Note that 'float' and 'double' are stored as 'float32' and 'float64'
-    assert ds.properties.data_layers["seg1"].element_class == "float32"
-    assert ds.properties.data_layers["seg2"].element_class == "float32"
-    assert ds.properties.data_layers["seg3"].element_class == "float32"
-    assert ds.properties.data_layers["seg4"].element_class == "float64"
-    assert ds.properties.data_layers["seg5"].element_class == "float96"
+    assert ds.get_layer("seg1").dtype_per_layer == "float32"
+    assert ds.get_layer("seg2").dtype_per_layer == "float32"
+    assert ds.get_layer("seg3").dtype_per_layer == "float32"
+    assert ds.get_layer("seg4").dtype_per_layer == "float64"
+    assert ds.get_layer("seg5").dtype_per_layer == "float96"
 
 
 def test_writing_subset_of_compressed_data_multi_channel() -> None:
@@ -1159,6 +1136,10 @@ def test_add_symlink_layer() -> None:
     copytree(
         TESTDATA_DIR / "simple_wk_dataset", TESTOUTPUT_DIR / "simple_wk_dataset_copy"
     )
+    # Add an additional segmentation layer to the original dataset
+    Dataset(TESTOUTPUT_DIR / "simple_wk_dataset_copy").add_layer(
+        "segmentation", LayerCategories.SEGMENTATION_TYPE, largest_segment_id=999
+    )
 
     original_mag = (
         Dataset(TESTOUTPUT_DIR / "simple_wk_dataset_copy")
@@ -1167,15 +1148,23 @@ def test_add_symlink_layer() -> None:
     )
 
     ds = Dataset.create(TESTOUTPUT_DIR / "wk_dataset_with_symlink", scale=(1, 1, 1))
+    # symlink color layer
     symlink_layer = ds.add_symlink_layer(
         TESTOUTPUT_DIR / "simple_wk_dataset_copy" / "color"
+    )
+    # symlink segmentation layer
+    symlink_segmentation_layer = ds.add_symlink_layer(
+        TESTOUTPUT_DIR / "simple_wk_dataset_copy" / "segmentation"
     )
     mag = symlink_layer.get_mag("1")
 
     assert (TESTOUTPUT_DIR / "wk_dataset_with_symlink" / "color" / "1").exists()
+    assert (TESTOUTPUT_DIR / "wk_dataset_with_symlink" / "segmentation").exists()
 
-    assert len(ds.properties.data_layers) == 1
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 1
+    assert len(ds.layers) == 2
+    assert len(ds.get_layer("color").mags) == 1
+
+    assert cast(SegmentationLayer, symlink_segmentation_layer).largest_segment_id == 999
 
     # write data in symlink layer
     write_data = (np.random.rand(3, 10, 10, 10) * 255).astype(np.uint8)
@@ -1203,16 +1192,16 @@ def test_add_symlink_mag(tmp_path: Path) -> None:
         offset=(6, 6, 6), data=(np.random.rand(10, 20, 30) * 255).astype(np.uint8)
     )
 
-    assert tuple(layer.get_bounding_box().topleft) == (6, 6, 6)
-    assert tuple(layer.get_bounding_box().size) == (10, 20, 30)
+    assert tuple(layer.bounding_box.topleft) == (6, 6, 6)
+    assert tuple(layer.bounding_box.size) == (10, 20, 30)
 
     symlink_mag = layer.add_symlink_mag(tmp_path / "original" / "color" / "2")
 
     assert (tmp_path / "link" / "color" / "1").exists()
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 2
+    assert len(layer._properties.wkw_resolutions) == 2
 
-    assert tuple(layer.get_bounding_box().topleft) == (0, 0, 0)
-    assert tuple(layer.get_bounding_box().size) == (16, 26, 36)
+    assert tuple(layer.bounding_box.topleft) == (0, 0, 0)
+    assert tuple(layer.bounding_box.size) == (16, 26, 36)
 
     # Write data in symlink layer
     # Note: The written data is fully inside the bounding box of the original data.
@@ -1241,16 +1230,16 @@ def test_add_copy_mag(tmp_path: Path) -> None:
         offset=(6, 6, 6), data=(np.random.rand(10, 20, 30) * 255).astype(np.uint8)
     )
 
-    assert tuple(layer.get_bounding_box().topleft) == (6, 6, 6)
-    assert tuple(layer.get_bounding_box().size) == (10, 20, 30)
+    assert tuple(layer.bounding_box.topleft) == (6, 6, 6)
+    assert tuple(layer.bounding_box.size) == (10, 20, 30)
 
     copy_mag = layer.add_copy_mag(tmp_path / "original" / "color" / "2")
 
     assert (tmp_path / "link" / "color" / "1").exists()
-    assert len(ds.properties.data_layers["color"].wkw_magnifications) == 2
+    assert len(layer._properties.wkw_resolutions) == 2
 
-    assert tuple(layer.get_bounding_box().topleft) == (0, 0, 0)
-    assert tuple(layer.get_bounding_box().size) == (16, 26, 36)
+    assert tuple(layer.bounding_box.topleft) == (0, 0, 0)
+    assert tuple(layer.bounding_box.size) == (16, 26, 36)
 
     # Write data in copied layer
     write_data = (np.random.rand(5, 5, 5) * 255).astype(np.uint8)
@@ -1388,7 +1377,7 @@ def test_for_zipped_chunks() -> None:
         .get_or_add_mag("1", block_len=8, file_len=4)
     )
 
-    target_mag.layer.set_bounding_box(offset=(0, 0, 0), size=(256, 256, 256))
+    target_mag.layer.bounding_box = BoundingBox((0, 0, 0), (256, 256, 256))
     target_view = target_mag.get_view(size=(256, 256, 256))
 
     with get_executor_for_args(None) as executor:
@@ -1542,11 +1531,11 @@ def test_compression(tmp_path: Path) -> None:
 
 def test_dataset_view_configuration(tmp_path: Path) -> None:
     ds1 = Dataset.create(tmp_path, scale=(2, 2, 1))
-    default_view_configuration = ds1.get_view_configuration()
+    default_view_configuration = ds1.default_view_configuration
     assert default_view_configuration is None
 
-    ds1.set_view_configuration(DatasetViewConfiguration(four_bit=True))
-    default_view_configuration = ds1.get_view_configuration()
+    ds1.default_view_configuration = DatasetViewConfiguration(four_bit=True)
+    default_view_configuration = ds1.default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.four_bit == True
     assert default_view_configuration.interpolation == None
@@ -1558,21 +1547,22 @@ def test_dataset_view_configuration(tmp_path: Path) -> None:
     assert default_view_configuration.rotation == None
 
     # Test if only the set parameters are stored in the properties
-    assert ds1.properties.default_view_configuration == {"fourBit": True}
+    with open(ds1.path / PROPERTIES_FILE_NAME) as f:
+        properties = json.load(f)
+        assert properties["defaultViewConfiguration"] == {"fourBit": True}
 
-    ds1.set_view_configuration(
-        DatasetViewConfiguration(
-            four_bit=True,
-            interpolation=False,
-            render_missing_data_black=True,
-            loading_strategy="PROGRESSIVE_QUALITY",
-            segmentation_pattern_opacity=40,
-            zoom=0.1,
-            position=(12, 12, 12),
-            rotation=(1, 2, 3),
-        )
+    ds1.default_view_configuration = DatasetViewConfiguration(
+        four_bit=True,
+        interpolation=False,
+        render_missing_data_black=True,
+        loading_strategy="PROGRESSIVE_QUALITY",
+        segmentation_pattern_opacity=40,
+        zoom=0.1,
+        position=(12, 12, 12),
+        rotation=(1, 2, 3),
     )
-    default_view_configuration = ds1.get_view_configuration()
+
+    default_view_configuration = ds1.default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.four_bit == True
     assert default_view_configuration.interpolation == False
@@ -1585,7 +1575,7 @@ def test_dataset_view_configuration(tmp_path: Path) -> None:
 
     # Test if the data is persisted to disk
     ds2 = Dataset(tmp_path)
-    default_view_configuration = ds2.get_view_configuration()
+    default_view_configuration = ds2.default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.four_bit == True
     assert default_view_configuration.interpolation == False
@@ -1597,40 +1587,41 @@ def test_dataset_view_configuration(tmp_path: Path) -> None:
     assert default_view_configuration.rotation == (1, 2, 3)
 
     # Test camel case
-    view_configuration_dict = ds2.properties.default_view_configuration
-    assert view_configuration_dict is not None
-    for k in view_configuration_dict.keys():
-        assert _snake_to_camel_case(k) == k
+    with open(ds1.path / PROPERTIES_FILE_NAME) as f:
+        properties = json.load(f)
+        view_configuration_dict = properties["defaultViewConfiguration"]
+        for k in view_configuration_dict.keys():
+            assert _snake_to_camel_case(k) == k
 
 
 def test_layer_view_configuration(tmp_path: Path) -> None:
     ds1 = Dataset.create(tmp_path, scale=(2, 2, 1))
     layer1 = ds1.add_layer("color", LayerCategories.COLOR_TYPE)
-    default_view_configuration = layer1.get_view_configuration()
+    default_view_configuration = layer1.default_view_configuration
     assert default_view_configuration is None
 
-    layer1.set_view_configuration(LayerViewConfiguration(color=(255, 0, 0)))
-    default_view_configuration = layer1.get_view_configuration()
+    layer1.default_view_configuration = LayerViewConfiguration(color=(255, 0, 0))
+    default_view_configuration = layer1.default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.color == (255, 0, 0)
     assert default_view_configuration.alpha is None
     assert default_view_configuration.intensity_range is None
     assert default_view_configuration.is_inverted is None
     # Test if only the set parameters are stored in the properties
-    assert ds1.properties.data_layers["color"].default_view_configuration == {
-        "color": (255, 0, 0)
-    }
+    with open(ds1.path / PROPERTIES_FILE_NAME) as f:
+        properties = json.load(f)
+        assert properties["dataLayers"][0]["defaultViewConfiguration"] == {
+            "color": [255, 0, 0]
+        }
 
-    layer1.set_view_configuration(
-        LayerViewConfiguration(
-            color=(255, 0, 0),
-            alpha=1.0,
-            min=55.0,
-            intensity_range=(-12.3e1, 123),
-            is_inverted=True,
-        )
+    layer1.default_view_configuration = LayerViewConfiguration(
+        color=(255, 0, 0),
+        alpha=1.0,
+        min=55.0,
+        intensity_range=(-12.3e1, 123),
+        is_inverted=True,
     )
-    default_view_configuration = layer1.get_view_configuration()
+    default_view_configuration = layer1.default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.color == (255, 0, 0)
     assert default_view_configuration.alpha == 1.0
@@ -1640,7 +1631,7 @@ def test_layer_view_configuration(tmp_path: Path) -> None:
 
     # Test if the data is persisted to disk
     ds2 = Dataset(tmp_path)
-    default_view_configuration = ds2.get_layer("color").get_view_configuration()
+    default_view_configuration = ds2.get_layer("color").default_view_configuration
     assert default_view_configuration is not None
     assert default_view_configuration.color == (255, 0, 0)
     assert default_view_configuration.alpha == 1.0
@@ -1649,12 +1640,13 @@ def test_layer_view_configuration(tmp_path: Path) -> None:
     assert default_view_configuration.min == 55.0
 
     # Test camel case
-    view_configuration_dict = ds2.properties.data_layers[
-        "color"
-    ].default_view_configuration
-    assert view_configuration_dict is not None
-    for k in view_configuration_dict.keys():
-        assert _snake_to_camel_case(k) == k
+    with open(ds2.path / PROPERTIES_FILE_NAME) as f:
+        properties = json.load(f)
+        view_configuration_dict = properties["dataLayers"][0][
+            "defaultViewConfiguration"
+        ]
+        for k in view_configuration_dict.keys():
+            assert _snake_to_camel_case(k) == k
 
 
 def test_get_largest_segment_id(tmp_path: Path) -> None:
@@ -1734,12 +1726,19 @@ def test_add_copy_layer(tmp_path: Path) -> None:
     original_color_layer.add_mag(1).write(
         offset=(10, 20, 30), data=(np.random.rand(32, 64, 128) * 255).astype(np.uint8)
     )
+    other_ds.add_layer(
+        "segmentation", LayerCategories.SEGMENTATION_TYPE, largest_segment_id=999
+    )
 
     # Copies the "color" layer from a different dataset
     ds.add_copy_layer(tmp_path / "other_ds" / "color")
-    assert len(ds.layers) == 1
+    ds.add_copy_layer(tmp_path / "other_ds" / "segmentation")
+    assert len(ds.layers) == 2
+    assert (
+        cast(SegmentationLayer, ds.get_layer("segmentation")).largest_segment_id == 999
+    )
     color_layer = ds.get_layer("color")
-    assert color_layer.get_bounding_box() == BoundingBox(
+    assert color_layer.bounding_box == BoundingBox(
         topleft=(10, 20, 30), size=(32, 64, 128)
     )
     assert color_layer.mags.keys() == original_color_layer.mags.keys()
@@ -1766,14 +1765,48 @@ def test_rename_layer(tmp_path: Path) -> None:
     write_data = (np.random.rand(10, 20, 30) * 255).astype(np.uint8)
     mag.write(data=write_data)
 
-    layer.rename("color2")
+    layer.name = "color2"
 
     assert not (tmp_path / "ds" / "color").exists()
     assert (tmp_path / "ds" / "color2").exists()
-    assert "color2" in ds.properties.data_layers.keys()
-    assert "color2" == ds.properties.data_layers["color2"].name
+    assert len([l for l in ds._properties.data_layers if l.name == "color"]) == 0
+    assert len([l for l in ds._properties.data_layers if l.name == "color2"]) == 1
     assert "color2" in ds.layers.keys()
     assert "color" not in ds.layers.keys()
 
     # The "mag" object which was created before renaming the layer is still valid
     assert np.array_equal(mag.read()[0], write_data)
+
+
+def test_delete_layer_and_mag(tmp_path: Path) -> None:
+    ds = Dataset.create(tmp_path / "ds", scale=(1, 1, 1))
+    color_layer = ds.add_layer("color", LayerCategories.COLOR_TYPE)
+    color_layer.add_mag(1)
+    color_layer.add_mag(2)
+    ds.add_layer(
+        "segmentation", LayerCategories.SEGMENTATION_TYPE, largest_segment_id=999
+    )
+    assert "color" in ds.layers
+    assert "segmentation" in ds.layers
+    assert len([l for l in ds._properties.data_layers if l.name == "color"]) == 1
+    assert len([l for l in ds._properties.data_layers if l.name == "segmentation"]) == 1
+    assert len(color_layer._properties.wkw_resolutions) == 2
+
+    color_layer.delete_mag(1)
+    assert len(color_layer._properties.wkw_resolutions) == 1
+    assert (
+        len(
+            [
+                m
+                for m in color_layer._properties.wkw_resolutions
+                if Mag(m.resolution) == Mag(2)
+            ]
+        )
+        == 1
+    )
+
+    ds.delete_layer("color")
+    assert "color" not in ds.layers
+    assert "segmentation" in ds.layers
+    assert len([l for l in ds._properties.data_layers if l.name == "color"]) == 0
+    assert len([l for l in ds._properties.data_layers if l.name == "segmentation"]) == 1