+++ Changelog.md
@@ -10,7 +10,12 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 [Commits](https://github.com/scalableminds/webknossos-cuber/compare/v0.8.12...HEAD)
 
 ### Breaking Changes in Config & CLI
-
+- Replaced the old properties classes of the Dataset API with attr-classes.
+  - The `Layer.rename()` function is now replaced with the setter of `Layer.name`.
+  - The functions `Layer.get_view_configuration` and `Layer.set_view_conficuration` are replaced by the property `Layer.default_view_configuration`. (Same applies to `Dataset.get_view_configuration` and `Dataset.set_view_configuration`)
+  - Moved `LayerViewConfiguration` and `DatasetViewConfiguration` into `properties.py`
+  - Removed `Layer.set_bounding_box_offset` and `Layer.set_bounding_box_size`.
+  - Renamed `Layer.get_bounding_box()` to the property `Layer.bounding_box`. The method `Layer.set_bounding_box` is replaced with the setter of the property `Layer.bounding_box`.
 ### Added
 - The API documentation is now hosted on a publicwebpage. [#392](https://github.com/scalableminds/webknossos-cuber/pull/392)